Tried to improve the channel through which visualizer info is communicated from the mobsim to the otfvis.  This used to be a completely un-typed collection of variables, and life was made harder by the fact that something that started as "type" eventually became "state".  It now tries to use an interface called "AgentSnapshotInfo".
This touches the Byte channel through which internally in the otfvis information is transmitted.  Unfortunately, the *.mvi files use the same format.  In consequence, old files cannot be read with the new mechanism.  I attempted to do something against that but no guarantee that this works.  Let us know if there are problems, but we can't promise anything because nobody here understand otfvis completely after David Strippgen has left.

It is pretty clear that new *.mvi files will not work with old otfvis.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@10152 a0517364-e873-4167-b1c8-416daa5a6b91