protected core.mobsim.queuesim even more from outside access.  With the introduction of "Vis..." and "CapacityInformation...", this went surprisingly well. (can't say how much I am building on top of other people's work).
Surprisingly, I still found a larger number of references to core.mobsim.queuesim inside otfvis; I had (probably naively) assumed that all of those had been pointered to ptproject.qsim.  Overall, however, this indicates to me that otfvis is maybe less dependent on the type of qsim than we tend to think ... and again, this may be the result of the work of others (in particular, Michael Z. said something similar to me last week).

git-svn-id: file:///srv/local/data/cleanRepo/trunk@11446 a0517364-e873-4167-b1c8-416daa5a6b91