I tried to make the "PositionInfo" mechanics configurable ...
... since I wanted to be able to switch on/off an adaptive way to compute the orthogonal offset (for driving "on the right"). (*)

I did this by pulling apart the pure data class, and the construction mechanism which converts graph coordinates (link,distance) to Euclidean coordinates (x,y).  The construction mechanism is now called "AgentSnapshotInfoFactory"; "PositionInfo" is now the default implementation of "AgentSnapshotInfo".  The factory has static methods that do the past (non-configurable) behavior; and it is meant to also be instantiable (and thus configurable in a non-static way).

In the end, I ended up making (*) the non-configurable default behavior, so I did not go to the end with making the AgentSnapshotInfoFactory configurable.  But the mechanics is now there.  Since I think it is useful, I leave it in place.

===

In the process, I made Gregors PositionInfo a separate implementation of AgentSnapshotInfo rather than something inherited.  It essentially already was a separate implementation anyways. 

===

I also exchanged the sequence of the otfvis tests, since the QSim_MVI test depended on the LiveVis test.

Also made those tests use the "queue" instead of the "equiDist" method, since I think it gives more information.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@12293 a0517364-e873-4167-b1c8-416daa5a6b91