changed ActivityImpl.getDuration() to ActivityImpl.getMaximumDuration()
removed the "deprecated" annotation

added it to the interface (in the api)

renamed "calculateDuration" into "calculateSomeDuration", because I don't think that durations from/to midnight are the way in which we should do this

made "calculateSomeDuration" a static utility method (based on the api) and moved it to org.matsim.utils.deprecated.DeprecatedStaticMethod.  Could, in principle, be moved to playground.mfeil, since 95% of is applications are there, but would need to fix the maven dependencies for the remaining two uses of the method.


git-svn-id: file:///srv/local/data/cleanRepo/trunk@14319 a0517364-e873-4167-b1c8-416daa5a6b91