Repairing wrashid playground tests.  Reasons for failures:
(1) PlansScoring.getScoringFunctionForAgent(id) was not adapted to the recent revision of Events2Score and thus returned a null pointer.  I am a bit amazed that nobody complained; maybe not so many persons are using this functionality (which is deprecated anyway).  This fixed most of the test failures.
(2) chessConfig2.xml contained a deprecated location choice config parameter.  In my memory, this was deprecated a long time ago, but I may be wrong.  Also, I don't think that location choice is used at all in that test.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@22427 a0517364-e873-4167-b1c8-416daa5a6b91