In an act of desperation, I just made Controler.createRoutingAlgorithm() final.  This means that it is no longer possible to override this method.  Use setTripRouterFactory instead.
Reason why I did that is that, after recent refactorings, createRoutingAlgorithm only affects initialization, but not the ReRoute during the runs.  Result is that code overriding createRoutingAlgorithm will still look like it is working, but it may not do what the user intended.  I have seen places where it was implemented correctly in spite of this but I find it just too dangerous.

For classes that overrode createRoutingAlgorithm() I commented it out.  I also tried to do the following:
* throw runtime exceptions during calls to the constructor of such classes with a warning message
* also do this to derived classes

I may not have caught all instances correctly, but I think that doing nothing is even worse.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@24262 a0517364-e873-4167-b1c8-416daa5a6b91