change the way "pseudo simulation" is handled
rather than having an inner loop, which was a messy thing relating event
handlers, scoring and the like (scoring of joint activities was not
supported for instance), just have a strategymanager and a
mobsimfactory switching back and forth between "PSim" and "Physical"
modes.
This also allows to monitor PSim iterations as normal iterations in
stopwatch, score evolution and the rest.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@29259 a0517364-e873-4167-b1c8-416daa5a6b91