change oss.sonatype.org repo from http to https
the redirection was not handled correctly, making maven go completely
crazy. Somehow eclipse did not complain too much, but building on the
command line was a nightmare, and the build server is under heavy
thunderstorm since a while now, with roughly half the playgrounds not
compiling.
This fix, together with cleaning my local ~/.m2/ repo, solved the problem
on my machine.
Hopefully the same happens automatically on the build server. If not,
deleting the ~/.m2/ of the server manually should do the trick.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@29438 a0517364-e873-4167-b1c8-416daa5a6b91