further move Gauteng road pricing into roadpricing contrib; remove usingRoadPricing in the process
Reason for removing that switch: We found it hard to interpret.  Should a script "set" this switch, or rather "interpret" it, or ignore it?  For the Gauteng toll simulation, it had to be set "false" in order to make everything work correctly.  It is now gone; if you want to simulate a non-toll base case, recommendation is to use an empty toll file.  In that way, you can be confident that you do not get two different execution paths which may cause differences by themselves.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@30125 a0517364-e873-4167-b1c8-416daa5a6b91