no more IdImpl in org.matsim.*, except org.matsim.contrib.* and the Implementation itself.
Had to adapt quite a lot of code due to the usage of Id for org.matsim.matrices.Matrix. Changed Id to String in Matrix as matrices were used to refer to lots of different things, the fewest of them having an actual type (class) to be used as type in the new typed Ids.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@30352 a0517364-e873-4167-b1c8-416daa5a6b91