The scenario config group is now essentially gone.
The class is still there, but only to result in somewhat meaningful error messages if someone has these switches in his or her config group.

The behavior is now resonably consistently that files are read into the containers if the filenames are given, and otherwise not.

Reading the files into their containers does NOT switch on corresponding functionality; that needs to be done separately and often elsewhere (e.g. in the qsim to respect lanes and/or vehicles).

The "refactoring" was not exact but based on heuristics, sometimes only string exchange, sometimes using "if ( filename!=null )" as heuristic to replace the previous test.  This may not always work as intended.