remove NetworkChangeEventFactoryImpl and related
I removed the factory machinery around NetworkChangeEvent.  Reasons:

(1) The factory returned a class, not an interface.  That makes replacing it with other implementations difficult; at best, could extend the returned class (but we don't like this), or assume that it will be replaced by an interface later.
(2) It is a short-lived "parameter" object that is only used to insert information elsewhere and then goes away.  Not very probable that this needs to exist with different interpretations.
(3) The syntax in the code was not consistent: Although it was possible to get it out of the network factory (which would have been the right way to go if ever accepting an alternative implementation), most places in the code in fact generated an ad-hoc instance of NetworkChangeEventFactory and then created NetworkChangeEvent instances from it.  This is, however, not more flexible than using the constructor right away.

One might have moved forward, creating all instances out of the NetworkFactory.  But given (1) and (2), that seemed to be a bit heavyweight even for MATSim standards :-).