Removing agarwalamit and benjamin playgrounds
Here is the email I wrote, explaining the situation, and also pointing to versions that still work:

Dear Amit, Benjamin,

The two playgrounds "agarwalamit" and "benjamin" have been unmaintained for so long that we are unfortunately unable to rescue them.

The problem roughly was that compile errors were fixed in your playgrounds, but failing tests were not. In consequence, it is no longer possible to find out at which point the results of the regression tests started failing.  Since these are all "content" tests (expecting some events file, or expecting a certain number of vehicles on a link), we would need to go through the failing tests, understand them, and then possibly fix them.  I don't think that we would be able to do this even if we had the resources.

An alternative would be to go back to some version that worked.  Since those two playgrounds actually did not change that much, one could maybe

* go back to Amit's last vsp-playground commit (c837c60d986f9f1d8bb2533d8fd9aa494193ee7a on 27/oct/2018)

* find a matsim commit with which vsp/agarwalamit and vsp/benjamin compile and test (I think that 3dedf36c56ca52b2d5fc39baf21419fbc29c2ed2 works)

* replay the matsim changes by slowing incrementing matsim (or bisection searching), fixing compilation and test errors in "agarwalamit" and "benjamin" as one goes

I just tried that; it seems possible in principle, but one needs to run the regression tests in "agarwalamit", "benjamin" and "vsp" many many times, and that takes time.

---

It is also not possible to make them depend on an old version of matsim, since with that I can in particular not set the dependency on the "vsp" playground to some older version.

---

From a more general perspective, I would assume that all the important material is somewhere in contribs, or possibly in the central "vsp" playground (which just became a contrib).

---

I would thus remove your two playgrounds, noting that they were still working with the above commit hashes. Please let me know if you are violently against this; we would then revert the commit.  In that case, you would need to fix those tests very soon.  Maybe you yourself would be able to fix the tests by looking at them.

It might be possible to only rescue one or the other playground, but for the time being agarwalamit depends on benjamin so I cannot treat them separately.

Very sorry, all the best

Kai